Many small tweaks for quality.
Dithering is now in screen space; this may be a problem on some models,
we'll see. Borders are drawn differently to ensure a roughly-consistent
thickness around the shrunken iso models. Letters have been removed from
most wargame voxel models, e.g. Truck_P doesn't have a S behind it
(which stood for Supply a long time ago), and Tank_S doesn't have a T
for Tank on its back anymore (but does have an up arrow instead since it
specifically is anti-air). Flamethrower has flaming torches above the
pilot since it's meant to destroy vegetation, as well as anyone in it,
and wanton torching does the job easily. There's some tweaks to simple
terrain to ensure it has a border.