I think that's the last of the changes for Above
A few more things may need to be added for non-wargame sprites, like the
new natures table that determines what gets a slope drawn and what color
it should be. Oher that that, the graphics just look a lot better, and
seem to be produced by the tool a fair bit more rapidly.