Writing to a fresh folder prevents overwrite bugs
This may require some code to delete the large folders entirely, but if
an old folder of images had different (larger, after compression) file
sizes, then it looks like the files that get written out only overwrite
some of those files, and leave the rest as garbage. This is OK for
libpng to read, since it is tolerant of poorly-formed PNGs, but other,
more spec-conformant libs may crash if they encounter data after a PNG
image's last chunk. The workaround seems to be writing to a new folder
if the rendered result ever changes.